requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
streamlit==1.37.0
langchain==0.2.14
langchain-community==0.2.12
langchain-huggingface==0.0.3
langchain-chroma==0.1.2
sentence-transformers==3.0.1
transformers==4.44.0
torch>=2.2.0
chromadb==0.5.5
google-generativeai==0.7.2
pypdf==4.3.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.4.0
unstructured==0.15.0
unstructured-inference==0.7.36
pdfminer.six==20240706
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.4
numpy>=1.26